Fixed/Moved UpdateCourseKeyView to API directory. ECOM-6692
