+++ course_discovery/apps/publisher/api/views.py
@@ -2,9 +2,11 @@ from rest_framework.generics import UpdateAPIView, ListAPIView, get_object_or_40
 from rest_framework.permissions import IsAuthenticated
 
 from course_discovery.apps.core.models import User
-from course_discovery.apps.publisher.models import CourseUserRole, OrganizationExtension
+from course_discovery.apps.publisher.models import CourseUserRole, OrganizationExtension, CourseRun
 from course_discovery.apps.publisher.api.permissions import CanViewAssociatedCourse, InternalUserPermission
-from course_discovery.apps.publisher.api.serializers import CourseUserRoleSerializer, GroupUserSerializer
+from course_discovery.apps.publisher.api.serializers import (
+    CourseUserRoleSerializer, GroupUserSerializer, UpdateCourseKeySerializer
+)
 
 
 class CourseRoleAssignmentView(UpdateAPIView):
@@ -21,3 +23,9 @@ class OrganizationGroupUserView(ListAPIView):
         org_extension = get_object_or_404(OrganizationExtension, organization=self.kwargs.get('pk'))
         queryset = User.objects.filter(groups__name=org_extension.group)
         return queryset
+
+
+class UpdateCourseKeyView(UpdateAPIView):
+    permission_classes = (IsAuthenticated, InternalUserPermission,)
+    queryset = CourseRun.objects.all()
+    serializer_class = UpdateCourseKeySerializer
